{{cookiecutter.project_slug}}/backend/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "{{cookiecutter.version}}"
description = "{{cookiecutter.short_description}}"
authors = ["{{cookiecutter.full_name}} <{{cookiecutter.email}}>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.3"
Django = "^4.1"
djangorestframework = "^3.13.1"
redis = "^4.3.4"
celery = "^5.2.7"
flake8 = "^5.0.4"
black = "^22.6.0"
django-environ = "^0.9.0"
gunicorn = "^20.1.0"
firebase-admin = "^5.2.0"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
